#198 Move Savebutton and alerts to new component
